Major refactoring, added DEoptim class with new abstractions
With this commit, we are explicitly conceptualising:
  1. Problem
  2. Objective
  3. Strategy
  4. Population
  5. Diagnostics

In particular, DEoptim class now consumes the instances of above
mentioned classes and evolves accordingly. Strategy is an interface
and comes with a simple implementation called SimpleStrategy. This
allows programmers to implement their own strategies.

Furthermore, Diagnostics class allows adding listeners to the following
events:
  1. evolutionStarted
  2. iterationStarted
  3. iterationFinished
  4. evolutionFinished